Update Spec and README for round 3; fix mypy nits
Docs: role-aware targets, weekend multiplier, capacity advisories, solver
budget control, and the upload/retry/feedback fixes recorded in the Spec
progress log and README (auto-target description, weekend definition,
new pre-solve advisories section, verification guarantees, changelog).

mypy: SchedulerSolver.solve accepts float seconds; schedule_quality's
role-range helper takes plain dicts instead of dynamic TypedDict keys.

+++ model/optimiser.py
@@ -669,7 +669,7 @@ class SchedulerSolver:
 
         self.model.Minimize(sum(terms))
 
-    def solve(self, time_limit_sec: int | None = None):
+    def solve(self, time_limit_sec: float | None = None):
         solver = cp_model.CpSolver()
         if not hasattr(solver, "OPTIMAL"):
             solver.OPTIMAL = getattr(cp_model, "OPTIMAL", 0)
